blind-assistant-app/backend/requirements.txt
fastapi
uvicorn
pyngrok
nest-asyncio
python-multipart
pillow
transformers
accelerate
qwen-vl-utils
torch
torchvision